@@ -2,11 +2,12 @@
 name = autorad
 version = 0.1.2
 author = Piotr Woznicki
-author_email = piotrekwoznicki@gmail.com
+author_email = piotr.a.woznicki@gmail.com
 packages = ['autorad']
 description = Radiomics-related modules for extraction and experimenting
 long_description = file:README.md
 long_description_content_type = text/markdown
+license = Apache 2.0
 url = https://github.com/pwoznicki/AutoRadiomics
 project_urls =
    Bug Tracker = https://github.com/pwoznicki/AutoRadiomics
@@ -15,7 +16,7 @@ classifiers =
    Operating System :: OS Independent
 
 [options]
-python_requires = >=3.8
+python_requires = >=3.10
 
 [options.packages.find]
 where = autorad
