@@ -3,21 +3,30 @@ from pathlib import Path
 
 import pandas as pd
 import pytest
+from hypothesis import settings
 
 from autorad.config import config
+from autorad.data.dataset import FeatureDataset
+
+settings.register_profile("fast", max_examples=2)
+settings.register_profile("slow", max_examples=10)
+
+prostate_root = Path(config.TEST_DATA_DIR) / "nifti" / "prostate"
+prostate_data = {
+    "img": prostate_root / "img.nii.gz",
+    "seg": prostate_root / "seg_one_label.nii.gz",
+    "seg_two_labels": prostate_root / "seg_two_labels.nii.gz",
+    "empty_seg": prostate_root / "seg_empty.nii.gz",
+}
 
 
 @pytest.fixture
 def small_paths_df():
-    nifti_dir = Path(config.TEST_DATA_DIR) / "nifti"
-    img_path = nifti_dir / "img.nii.gz"
-    one_label_mask_path = nifti_dir / "seg_one_label.nii.gz"
-    multi_label_mask_path = nifti_dir / "seg_multi_label.nii.gz"
     paths_df = pd.DataFrame(
         {
-            "ID": [0, 1],
-            "img": [img_path, img_path],
-            "seg": [one_label_mask_path, multi_label_mask_path],
+            "ID": ["case_1_single_label", "case_2_two_labels"],
+            "img": [prostate_data["img"], prostate_data["img"]],
+            "seg": [prostate_data["seg"], prostate_data["seg_two_labels"]],
         }
     )
     return paths_df
@@ -52,13 +61,30 @@ def multiclass_df():
 
 
 @pytest.fixture
-def test_dfs():
-    return [binary_df, multiclass_df]
+def feature_dataset(df):
+    return FeatureDataset(
+        dataframe=df,
+        ID_colname="ID",
+        target="Label",
+    )
+
+
+def feature_df():
+    return pd.DataFrame(
+        {
+            "ID": [str(i) for i in range(100)],
+            "Feature1": [100 * i for i in range(100)],
+            "Feature2": [i % 2 for i in range(100)],
+            "Label": [i % 2 for i in range(100)],
+        }
+    )
 
 
 class Helpers:
-    # for common utils, following advice from
-    # https://stackoverflow.com/questions/33508060/create-and-import-helper-functions-in-tests-without-creating-packages-in-test-di
+    """For common utils, following advice from
+    https://stackoverflow.com/questions/33508060
+    """
+
     @staticmethod
     def tmp_dir():
         dirpath = tempfile.mkdtemp()
